[20/21] Move code from app to controllers
Pagination, get_pagination
