+++ src/azul/service/repository_controller.py
@@ -4,20 +4,27 @@ from collections.abc import (
 from functools import (
     partial,
 )
+import json
 import logging
 from typing import (
     Any,
     cast,
 )
 
+import attr
 from chalice import (
     BadRequestError,
+    ChaliceViewError,
     NotFoundError,
     Response,
+)
+from furl import (
+    furl,
 )
 
 from azul import (
     CatalogName,
+    R,
     cache,
     cached_property,
     iif,
@@ -65,6 +72,50 @@ log = logging.getLogger(__name__)
 
 
 class RepositoryController(ServiceController):
+    @attr.s(kw_only=True, auto_attribs=True, frozen=True)
+    class Pagination(Pagination):
+        self_url: furl
+
+        def link(self, *, previous: bool, **params: str) -> furl | None:
+            search_key = self.search_before if previous else self.search_after
+            if search_key is None:
+                return None
+            else:
+                before_or_after = 'before' if previous else 'after'
+                params = {
+                    **params,
+                    f'search_{before_or_after}': json.dumps(search_key),
+                    'sort': self.sort,
+                    'order': self.order,
+                    'size': self.size
+                }
+            return furl(url=self.self_url, args=params)
+
+    def get_pagination(self, entity_type: str) -> Pagination:
+        default_sorting = self.app.metadata_plugin.exposed_indices[entity_type]
+        params = self.app.current_request.query_params or {}
+        sb, sa = params.get('search_before'), params.get('search_after')
+        if sb is None:
+            if sa is not None:
+                sa = tuple(json.loads(sa))
+        else:
+            if sa is None:
+                sb = tuple(json.loads(sb))
+            else:
+                raise BadRequestError('Only one of search_after or search_before may be set')
+        try:
+            return self.Pagination(order=params.get('order', default_sorting.order),
+                                   size=int(params.get('size', '10')),
+                                   sort=params.get('sort', default_sorting.field_name),
+                                   search_before=sb,
+                                   search_after=sa,
+                                   self_url=self.app.self_url)
+        except AssertionError as e:
+            if R.caused(e):
+                raise R.propagate(e, ChaliceViewError)
+            else:
+                raise
+
     min_page_size = 1
 
     generic_object_spec = schema.object(additionalProperties=True)
@@ -326,7 +377,7 @@ class RepositoryController(ServiceController):
                                    entity_type=entity_type,
                                    item_id=entity_id,
                                    filters=query_params.get('filters'),
-                                   pagination=self.app.get_pagination(entity_type),
+                                   pagination=self.get_pagination(entity_type),
                                    authentication=request.authentication)
             return '' if request.method == 'HEAD' else response
